Move redact_url helper to utils module
Extract the URL redaction helper from session.py into a dedicated
utils module and cover it with a standalone test_utils.py suite,
as requested in the PR review.
